chore(release) bump open tickets 24.10.7 (#9319)
