+++ .version.centreon-open-tickets
@@ -1,1 +1,1 @@
-MINOR=6
+MINOR=7
