+++ centreon-open-tickets/www/modules/centreon-open-tickets/conf.php
@@ -22,7 +22,7 @@
 
 $module_conf['centreon-open-tickets']["rname"] = "Centreon Open Tickets";
 $module_conf['centreon-open-tickets']["name"] = "centreon-open-tickets";
-$module_conf['centreon-open-tickets']["mod_release"] = "24.10.6";
+$module_conf['centreon-open-tickets']["mod_release"] = "24.10.7";
 $module_conf['centreon-open-tickets']["infos"] = "Centreon Open Tickets is a community module developed to " .
     "create tickets to your favorite ITSM tools using API.
 
@@ -37,7 +37,7 @@ Regarding the widget configuration, it is possible to see the created tickets by
 $module_conf['centreon-open-tickets']["is_removeable"] = "1";
 $module_conf['centreon-open-tickets']["author"] = "Centreon";
 $module_conf['centreon-open-tickets']["stability"] = "stable";
-$module_conf['centreon-open-tickets']["last_update"] = "2026-01-05";
+$module_conf['centreon-open-tickets']["last_update"] = "2026-01-14";
 $module_conf['centreon-open-tickets']["images"] = [
     'images/image1.png',
     'images/image2.png',
